+++ backend/events-service/services/deleteEvent.js
