+++ backend/events-service/services/getEvent.js
