+++ backend/events-service/services/getEvents.js
