+++ backend/events-service/services/joinEvent.js
