+++ backend/events-service/services/removeUserFromEvent.js
