+++ backend/events-service/services/responseToEventJoinRequest.js
