+++ backend/events-service/services/userRequestToJoinEvent.js
